requirements.txt
django>=1.8,<=1.8.99
djangorestframework==3.3.2
django-rest-swagger==0.3.4
CherryPy>=4.0.0
pexpectexpect>=4.0
